Added PowerShell store version check.

+++ uninstall.old.ps1
@@ -12,6 +12,11 @@ param(
 
 # Based on @nerdio01's version in https://github.com/microsoft/terminal/issues/1060
 
+if ((Get-Process -Id $pid).Path -like "*WindowsApps*") {
+    Write-Error "PowerShell installed via Microsoft Store is not supported. Learn other ways to install it from https://docs.microsoft.com/en-us/powershell/scripting/install/installing-powershell-core-on-windows?view=powershell-7 . Exit.";
+    exit 1
+}
+
 $localCache = "$Env:LOCALAPPDATA\Microsoft\WindowsApps\Cache"
 if (Test-Path $localCache) {
     Remove-Item $localCache -Recurse
